+++ pyvistaqt/plotting.py
@@ -62,7 +62,6 @@ from qtpy.QtWidgets import (
     QFrame,
     QGestureEvent,
     QGridLayout,
-    QMenuBar,
     QToolBar,
     QVBoxLayout,
     QWidget,
@@ -72,7 +71,13 @@ from vtk.qt.QVTKRenderWindowInteractor import QVTKRenderWindowInteractor
 from .counter import Counter
 from .dialog import FileDialog, ScaleAxesDialog
 from .editor import Editor
-from .utils import _check_type
+from .utils import (
+    _check_type,
+    _setup_ipython,
+    _setup_application,
+    _setup_off_screen,
+    _create_menu_bar,
+)
 from .window import MainWindow
 
 if scooby.in_ipython():  # pragma: no cover
@@ -956,50 +961,3 @@ class MultiPlotter:
         row, col = idx
         self._plotter = self._plotters[row * self._ncols + col]
         return self._plotter
-
-
-def _create_menu_bar(parent: Any) -> QMenuBar:
-    """Create a menu bar.
-
-    The menu bar is expected to behave consistently
-    for every operating system since `setNativeMenuBar(False)`
-    is called by default and therefore lifetime and ownership can
-    be tested.
-    """
-    menu_bar = QMenuBar(parent=parent)
-    menu_bar.setNativeMenuBar(False)
-    if parent is not None:
-        parent.setMenuBar(menu_bar)
-    return menu_bar
-
-
-def _setup_ipython(ipython: Any = None) -> Any:
-    # ipython magic
-    if scooby.in_ipython():  # pragma: no cover
-        # pylint: disable=import-outside-toplevel
-        from IPython import get_ipython
-
-        ipython = get_ipython()
-        ipython.magic("gui qt")
-
-        # pylint: disable=redefined-outer-name
-        # pylint: disable=import-outside-toplevel
-        from IPython.external.qt_for_kernel import QtGui
-
-        QtGui.QApplication.instance()
-    return ipython
-
-
-def _setup_application(app: Optional[QApplication] = None) -> QApplication:
-    # run within python
-    if app is None:
-        app = QApplication.instance()
-        if not app:  # pragma: no cover
-            app = QApplication(["PyVista"])
-    return app
-
-
-def _setup_off_screen(off_screen: Optional[bool] = None) -> bool:
-    if off_screen is None:
-        off_screen = pyvista.OFF_SCREEN
-    return off_screen
